+++ versions/meta.py
@@ -5,11 +5,11 @@ from typing_extensions import Final
 import versions
 from versions.segments import PreTag
 from versions.version import Version
-from versions.versioned import get_version_unchecked
+from versions.versioned import get_version
 
 __all__ = ("version_info", "python_version_info")
 
-version_info = get_version_unchecked(versions)  # type: ignore
+version_info = get_version(versions)  # type: ignore
 """The library version represented as a [`Version`][versions.version.Version]."""
 
 FINAL: Final[str] = "final"